// ssyweb/SSYNEW/src/app/models/rlobankdetails.model.ts
export class RloBankDetails {
    rloOfficeBankId: number;
    rloOfficeId: number;
    bankName: string;
    bankBranch: string;
    addressLine1: string;
    addressLine2: string;
    ifsc: string;
    contactName: string;
    statusId: number;
    createdBy: null;
    creadtedDate: null;
    updatedBy: null;
    upadtedDate: null;
    openBalance: number;
    openBalanceDate: string;
    bankAccountNo: string;
}