// ssyweb/SSYNEW/src/app/models/paymentProcessDetails.model.ts
export class PaymentProcessMasterData {
    paymentProcessingID: number;
    fundRequestId: any;
    rloOfficeId: number;
    paymentProcessingDate: any;
    paymentReleaseDate: any;
    statusId: any;
    createdBy: any;
    creadtedDate: any;
    updatedBy: any;
    upadtedDate: any;
    rloOfficeName: string;
    paymentProcessingClaims: any;
    statusName: string;
    paymentProcessingDateString: any;
    paymentReleaseDateString: any;
    fundRequestNo: any;
}