// ssyweb/SSYNEW/src/app/models/IFMSDetails.model.ts
export class IFMSDetails {
    beneficiaryName: string;
    accountNumber: string;
    iFSCCode: string;
    accountType: string;
    beneficiaryType: string;
    mobileNumber: string;
    aadharNumber: string;
    address: string;
    emailId: string;
    amount: string;
    ssin: string;
    claimType: string;
}